set econ's and global forwaard pd into parts and generate fac.jld file

+++ ProdCode/Industry_Level/cust_data.jl
@@ -1,13 +1,13 @@
 function cust_data(data, dataEndMth, custMth, dateVctr = [])
-    # data, dataEndMth, custMth = dataFlatMth, dataEndMth, options["startMth"]
+    # data, dataEndMth, custMth = dataMtrxPD, endMth, options["startMth"]
      dataendmth = Dict()
-     dataendmth["year"] =  Int64(fld(dataEndMth, 100))
-     dataendmth["month"] =  Int64(dataEndMth - dataendmth["year"]*100)
+     dataendmth["year"] =  fld(dataEndMth, 100)
+     dataendmth["month"] =  dataEndMth - dataendmth["year"]*100
 
-     custMth_year = Int64(fld(custMth, 100))
-     custMth_month = Int64(custMth - custMth_year*100)
+     custMth_year = fld(custMth, 100)
+     custMth_month = custMth - custMth_year*100
 
-     validMth = (dataendmth["year"] - custMth_year)*12 + dataendmth["month"] - custMth_month + 1
+     validMth = Int((dataendmth["year"] - custMth_year)*12 + dataendmth["month"] - custMth_month + 1)
 
      totalMth = size(data, 1)
      data = data[(totalMth - validMth + 1):end, :, :]
